+++ lib/src/locale/code_output.dart
@@ -5,9 +5,11 @@ class CodeOutput {
     required this.classArgumentCode,
     required this.classBodyCode,
     required this.externalCode,
+    this.factoryArgumentCode,
   });
 
   final String classArgumentCode;
   final String classBodyCode;
   final String externalCode;
+  final String? factoryArgumentCode;
 }