+++ lib/src/locale/localization_unit.dart
@@ -52,6 +52,7 @@ class GenderUnit extends LocalizationUnit {
     required this.parents,
   });
 
+  /// ? name of the field
   @override
   final String key;