+++ lib/src/locale/unit_to_interface_generator/gender_unit_to_interface_generator.dart
@@ -7,21 +7,19 @@ import '../localization_unit.dart';
 CodeOutput genderUnitToInterface(GenderUnit unit, {bool useThisKeyword = true}) {
   final Set<String> arguments = extractArguments(genderValueToString(unit.value)).toSet();
   String parentClassName = unit.parents.map(capitalize).join();
+
   if (parentClassName.isNotEmpty) {
     parentClassName = '$parentClassName.';
   }
 
   if (arguments.isEmpty) {
-    return CodeOutput(
-      classArgumentCode:
-          '${useThisKeyword ? 'required this.' : ''}${unit.key}${useThisKeyword ? '' : ':'}${useThisKeyword ? '' : ' $parentClassName\$${unit.key}'},',
-      classBodyCode: '''
-${unit.value.description != null ? '/// ${unit.value.description}' : ''}
-final String Function(Gender gender) ${unit.key};
-''',
-      externalCode: '',
+    return _empty(
+      unit: unit,
+      parentClassName: parentClassName,
+      useThisKeyword: useThisKeyword,
     );
   }
+
   String functionArguments = arguments.map((String arg) => 'required String $arg').join(', ');
   functionArguments = '(Gender gender, {$functionArguments})';
 
@@ -31,6 +29,34 @@ final String Function(Gender gender) ${unit.key};
     classBodyCode: '''
 ${unit.value.description != null ? '/// ${unit.value.description}' : ''}
 final String Function$functionArguments ${unit.key};
+''',
+    externalCode: '',
+  );
+}
+
+CodeOutput _empty({
+  required GenderUnit unit,
+  required String parentClassName,
+  required bool useThisKeyword,
+}) {
+  return CodeOutput(
+    classArgumentCode: [
+      if (useThisKeyword) 'required this.',
+      unit.key,
+      if (useThisKeyword == false) ': $parentClassName\$${unit.key}',
+    ].join(),
+    factoryArgumentCode: '''
+${unit.key}: (Gender gender) => Intl.gender(
+        gender.name,
+        name: r$qt${unit.key}$qt,
+        female: json[r$qt${unit.key}$qt]['female'].toString(), 
+        male: json[r$qt${unit.key}$qt]['male'].toString(), 
+        other: json[r$qt${unit.key}$qt]['other'].toString(),
+      ),
+''',
+    classBodyCode: '''
+${unit.value.description != null ? '/// ${unit.value.description}' : ''}
+final String Function(Gender gender) ${unit.key};
 ''',
     externalCode: '',
   );