+++ lib/src/template/class_factory_template.dart
@@ -1,0 +1,12 @@
+String classFactoryBeginningTemplate({required String className}) {
+  return '''
+factory $className.fromJson(Map<String, dynamic> json) {
+    return $className(
+''';
+}
+
+String classFactoryEndTemplate() {
+  return '''
+    );
+  }''';
+}