+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Before release:
+- return back all example localizations
+- unmark example as excluded
+- generate the lib for example
+- ???
+
 ### 1.1.0 - Added localization providers for loading localizations remotely
 ### 1.0.3 - Hotfix
 ### 1.0.2 - Screening of variable names in dynamic objects