+++ CHANGELOG.md
@@ -2,6 +2,9 @@ Before release:
 - return back all example localizations
 - unmark example as excluded
 - generate the lib for example
+- handle country code: en_US / en_UK / en_CA
+- merge different translations
+  - can use one as a primary, and change fields in another, only those, what presents (including for a remote source)
 - ???
 
 ### 1.1.0 - Added localization providers for loading localizations remotely